+++ src/main/java/com/example/spot/member/application/legacy/MemberService.java
@@ -18,12 +18,7 @@ public interface MemberService extends UserDetailsService {
     // 테스트 용 멤버 생성
     MemberResponseDTO.MemberTestDTO testMember(MemberInfoListDTO memberInfoListDTO);
 
-    MemberResponseDTO.SocialLoginSignInDTO signUpByKAKAO(String code) throws JsonProcessingException;
 
-    MemberResponseDTO.SocialLoginSignInDTO signUpByKAKAOForTest(String code)
-        throws JsonProcessingException;
-
-    void redirectURL() throws IOException;
 
     Member findMemberByEmail(String email);
